s2: Pin haxxed Goodix blobs
Our Goodix blobs were compiled for Android 6.0.1, which wasn't issue in Android versions <=9, but now started to be one.
We often get stack corruption even though stack isn't really corrupted, sometimes mutex is locked with nullptr and a lot of stuff like that.

Pinning it now because I forgot to do it before